requirements.txt
requests>=2.32.3
trafilatura>=1.8.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
pypdf>=5.0.1
python-docx>=1.1.2
pyyaml>=6.0.2
tqdm>=4.66.4
chardet>=5.2.0
fake-useragent>=1.4.0
urllib3>=2.2.2
sentence-transformers>=3.0.1
# For CUDA support, install PyTorch separately:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# torch>=2.3.1
faiss-cpu>=1.8.0.post1
numpy>=1.26.4
rank-bm25>=0.2.2
transformers>=4.44.2

